+++ .env
@@ -6,3 +6,5 @@ DEBUG=True
 YAGPT_CATALOG_ID=b1g659p7ld8o2k31g37c
 YAGPT_API_KEY=AQVN0ZRFbyqk97EWUy0KAG9jzqs6oUdcXrNYjYJ-
 YAGPT_TIME_TO_SLEEP=2
+
+USER_DOCUMENTS_STORE=user_docs
