+++ data/config.py
@@ -1,3 +1,4 @@
+import pathlib
 from dataclasses import dataclass
 
 from environs import Env
@@ -21,11 +22,17 @@ class YaGPT:
 
 
 @dataclass
+class UserFileStorage:
+    path: pathlib.Path
+
+
+@dataclass
 class Config:
     tg_bot: TgBot
     db: DbConfig
     ya_gpt: YaGPT
     debug: bool
+    user_file_storage: UserFileStorage
 
     def __getitem__(self, item: str):
         if hasattr(self, item):
@@ -48,9 +55,12 @@ def load_config(path: str = None):
             postgres_uri=env.str("POSTGRES_URI")
         ),
         ya_gpt=YaGPT(
-          catalog_id=env.str("YAGPT_CATALOG_ID"),
-          api_key=env.str("YAGPT_API_KEY"),
-          time_to_sleep=env.float("YAGPT_TIME_TO_SLEEP")
+            catalog_id=env.str("YAGPT_CATALOG_ID"),
+            api_key=env.str("YAGPT_API_KEY"),
+            time_to_sleep=env.float("YAGPT_TIME_TO_SLEEP")
         ),
         debug=env.bool("DEBUG", False),
+        user_file_storage=UserFileStorage(
+            path=pathlib.Path(env.str("USER_DOCUMENTS_STORE"))
+        )
     )
